feat: 增加账单精度

+++ src/main/java/com/chenbitao/noodle_shop/domain/model/MenuItem.java
@@ -1,5 +1,7 @@
 package com.chenbitao.noodle_shop.domain.model;
 
+import java.math.BigDecimal;
+
 public enum MenuItem {
     /** 大碗牛肉面 */
     BEEF_NOODLE_LARGE(18),
@@ -22,9 +24,9 @@ public enum MenuItem {
     /** 套餐2 */
     SET_MEAL_2(40);
 
-    private final int price;
+    private final BigDecimal price;
 
-    MenuItem(int price) { this.price = price; }
+    MenuItem(int price) { this.price = BigDecimal.valueOf(price); }
 
-    public int getPrice() { return price; }
+    public BigDecimal getPrice() { return price; }
 }
